Rename parallel_iterator -> iterator.
