+++ src/iterator.rs
