+++ src/lib.rs
@@ -15,9 +15,9 @@
 pub mod color;
 pub mod data_provider;
 pub mod errors;
+pub mod iterator;
 pub mod math;
 pub mod octree;
-pub mod parallel_iterator;
 pub mod read_write;
 pub mod s2_cells;
