+++ src/s2_cells/mod.rs
@@ -1,7 +1,7 @@
 use crate::data_provider::DataProvider;
 use crate::errors::*;
+use crate::iterator::{FilteredPointsIterator, PointCloud, PointLocation, PointQuery};
 use crate::math::{Isometry3, Obb};
-use crate::parallel_iterator::{FilteredPointsIterator, PointCloud, PointLocation, PointQuery};
 use crate::proto;
 use crate::read_write::{Encoding, PointIterator};
 use crate::{AttributeDataType, CURRENT_VERSION};
